+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Text of checkbox "Mailversand aktivieren" didn't change anymore #1954 !3358 @ToLei
 - Restricted access to email attachments !3345 @alex.simm
 - Stores, where you are a jumper are now displayed at the bottom of the stores dropdown menu !3364 @AntonBallmaier
+- Improved the calendar export !3378 @AntonBallmaier
 
 ## Bugfixes
 - Changed the headline for group applications #1916 !3353 @ToLei